backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
google-generativeai

